Renamed a few missed integration test classes
