+++ whois-commons/src/test/java/net/ripe/db/whois/common/aspects/RetryForAspectIntegrationTest.java
@@ -1,11 +1,9 @@
-package net.ripe.db.whois.db;
+package net.ripe.db.whois.common.aspects;
 
-import net.ripe.db.whois.common.aspects.RetryFor;
 import net.ripe.db.whois.common.support.AbstractDaoIntegrationTest;
 import org.junit.Before;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
 
 import java.io.FileNotFoundException;
 import java.io.IOException;
@@ -15,18 +13,13 @@ import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.fail;
 
-/**
- * Test to ensure that the Retry aspect is properly applied from whois-commons onto whois sub-modules.
- * <p/>
- * This test also fails running in your IDE if AspectJ is not configured correctly.
- */
-@ContextConfiguration(locations = {"classpath:applicationContext-whois-test.xml"})
-public class RetryForAspectTest extends AbstractDaoIntegrationTest {
+public class RetryForAspectIntegrationTest extends AbstractDaoIntegrationTest {
     static final int ATTEMPTS = 5;
 
     AtomicInteger attemptCounter;
 
-    @Autowired RetryForAspectOnClass retryForAspectOnClass;
+    @Autowired RetryForAspectMethod retryForAspectMethod;
+    @Autowired RetryForAspectType retryForAspectType;
 
     @Before
     public void setUp() throws Exception {
@@ -50,7 +43,7 @@ public class RetryForAspectTest extends AbstractDaoIntegrationTest {
 
     private void retryForAnnotatedMethod(final Exception e, final int expectedAttempts) throws Exception {
         try {
-            retryForAspectOnClass.incrementAndThrowException(attemptCounter, e);
+            retryForAspectMethod.incrementAndThrowException(attemptCounter, e);
             fail("Expected exception");
         } catch (Exception exc) {
             assertThat(e, is(exc));
@@ -76,18 +69,12 @@ public class RetryForAspectTest extends AbstractDaoIntegrationTest {
 
     private void retryForAnnotatedType(final Exception e, final int expectedAttempts) throws Exception {
         try {
-            incrementAndThrowException(attemptCounter, e);
+            retryForAspectType.incrementAndThrowException(attemptCounter, e);
             fail("Expected exception");
         } catch (Exception exc) {
             assertThat(e, is(exc));
         }
 
         assertThat(attemptCounter.get(), is(expectedAttempts));
-    }
-
-    @RetryFor(value = IOException.class, attempts = RetryForAspectTest.ATTEMPTS, intervalMs = 0)
-    public void incrementAndThrowException(final AtomicInteger counter, final Exception e) throws Exception {
-        counter.incrementAndGet();
-        throw e;
     }
 }
