+++ whois-commons/src/test/java/net/ripe/db/whois/common/aspects/RetryForAspectMethodImpl.java
@@ -6,7 +6,7 @@ import java.io.IOException;
 import java.util.concurrent.atomic.AtomicInteger;
 
 @Component
-@RetryFor(value = IOException.class, attempts = RetryForAspectTest.ATTEMPTS, intervalMs = 0)
+@RetryFor(value = IOException.class, attempts = RetryForAspectIntegrationTest.ATTEMPTS, intervalMs = 0)
 class RetryForAspectMethodImpl implements RetryForAspectMethod {
     @Override
     public void incrementAndThrowException(final AtomicInteger counter, final Exception e) throws Exception {
