+++ whois-commons/src/test/java/net/ripe/db/whois/common/aspects/RetryForAspectType.java
@@ -4,6 +4,6 @@ import java.io.IOException;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public interface RetryForAspectType {
-    @RetryFor(value = IOException.class, attempts = RetryForAspectTest.ATTEMPTS, intervalMs = 0)
+    @RetryFor(value = IOException.class, attempts = RetryForAspectIntegrationTest.ATTEMPTS, intervalMs = 0)
     void incrementAndThrowException(AtomicInteger counter, Exception e) throws Exception;
 }
