+++ whois-commons/src/test/java/net/ripe/db/whois/common/aspects/RetryForAspectTypeImpl.java
@@ -8,7 +8,7 @@ import java.util.concurrent.atomic.AtomicInteger;
 @Component
 class RetryForAspectTypeImpl implements RetryForAspectType {
     @Override
-    @RetryFor(value = IOException.class, attempts = RetryForAspectTest.ATTEMPTS, intervalMs = 0)
+    @RetryFor(value = IOException.class, attempts = RetryForAspectIntegrationTest.ATTEMPTS, intervalMs = 0)
     public void incrementAndThrowException(final AtomicInteger counter, final Exception e) throws Exception {
         counter.incrementAndGet();
         throw e;
