+++ whois-db/src/test/java/net/ripe/db/whois/db/RetryForAspectOnClass.java
@@ -7,7 +7,7 @@ import java.io.IOException;
 import java.util.concurrent.atomic.AtomicInteger;
 
 @Component
-@RetryFor(value = IOException.class, attempts = RetryForAspectTest.ATTEMPTS, intervalMs = 0)
+@RetryFor(value = IOException.class, attempts = RetryForAspectIntegrationTest.ATTEMPTS, intervalMs = 0)
 class RetryForAspectOnClass {
     public void incrementAndThrowException(final AtomicInteger counter, final Exception e) throws Exception {
         counter.incrementAndGet();
